Bring back `optional()` closing `}` and `)`
This required splitting the precedence rules for blocks and calls, as blocks need a combination of `prec.right()` + having the same prec level as general `_expression`s for this to work. In theory it seems like calls might just need to have 1 level higher than the prec level of `_expression`s, but right now it is set to 22.